Improved modular system, added necessary dependencies
