+++ requirements.txt
@@ -1,0 +1,5 @@
+lmgtfy==0.1.2
+lxml==3.5.0
+PyYAML==3.11
+requests==2.9.1
+wheel==0.24.0
